+++ blosc/blosc.c
@@ -91,43 +91,41 @@
 /* Synchronization variables */
 
 
-struct blosc_context {
-  int32_t compress;
-  /* 1 if we are doing compression 0 if decompress */
-
+struct blosc_context_s {
   const uint8_t* src;
+  /* The source buffer */
   uint8_t* dest;
-  /* The current pos in the destination buffer */
+  /* The destination buffer */
   uint8_t* header_flags;
-  /* Flags for header.  Currently booked:
-                            - 0: byte-shuffled?
-                            - 1: memcpy'ed?
-                            - 2: bit-shuffled? */
+  /* Flags for header */
   int32_t sourcesize;
-  /* Number of bytes in source buffer (or uncompressed bytes in compressed file) */
+  /* Number of bytes in source buffer */
   int32_t nblocks;
   /* Number of total blocks in buffer */
   int32_t leftover;
   /* Extra bytes at end of buffer */
   int32_t blocksize;
   /* Length of the block in bytes */
-  int32_t typesize;
-  /* Type size */
   int32_t num_output_bytes;
   /* Counter for the number of output bytes */
   int32_t destsize;
   /* Maximum size for destination buffer */
   uint8_t* bstarts;
   /* Start of the buffer past header info */
-  int32_t compcode;
+  uint8_t typesize;
+  /* Type size */
+  uint8_t compcode;
   /* Compressor code to use */
-  int clevel;
+  int8_t clevel;
   /* Compression level (1-9) */
+  uint8_t filtercode;
+  /* The code of the filter */
   schunk_header* schunk;
-  /* Super-chunk (if available) */
-
-  /* Cache for temporaries for serial operation */
+  /* Associated super-chunk (if available) */
   struct thread_context* serial_context;
+  /* Cache for temporaries for serial operation */
+  uint8_t compress;
+  /* 1 if we are doing compression 0 if decompress */
 
   /* Threading */
   int32_t nthreads;
@@ -152,7 +150,7 @@ struct blosc_context {
 };
 
 struct thread_context {
-  struct blosc_context* parent_context;
+  blosc_context* parent_context;
   int32_t tid;
   uint8_t* tmp;
   uint8_t* tmp2;
@@ -166,7 +164,7 @@ struct thread_context {
 };
 
 /* Global context for non-contextual API */
-static struct blosc_context* g_global_context;
+static blosc_context* g_global_context;
 static pthread_mutex_t global_comp_mutex;
 static int32_t g_compressor = BLOSC_BLOSCLZ;
 /* the compressor to use by default */
@@ -177,10 +175,10 @@ static schunk_header* g_schunk = NULL;   /* the pointer to super-chunk */
 
 
 /* Wrapped function to adjust the number of threads used by blosc */
-int blosc_set_nthreads_(struct blosc_context*);
+int blosc_set_nthreads_(blosc_context*);
 
 /* Releases the global threadpool */
-int blosc_release_threadpool(struct blosc_context* context);
+int blosc_release_threadpool(blosc_context* context);
 
 /* Macros for synchronization */
 
@@ -552,7 +550,7 @@ static int zstd_wrap_decompress(struct thread_context* thread_context,
 #endif /*  HAVE_ZSTD */
 
 /* Compute acceleration for blosclz */
-static int get_accel(const struct blosc_context* context) {
+static int get_accel(const blosc_context* context) {
   int32_t clevel = context->clevel;
   int32_t typesize = context->typesize;
 
@@ -579,7 +577,7 @@ static int blosc_c(struct thread_context* thread_context, int32_t blocksize,
                    int32_t leftoverblock, int32_t ntbytes, int32_t maxbytes,
                    const uint8_t* src, int offset, uint8_t* dest,
                    uint8_t* tmp, uint8_t* tmp2) {
-  struct blosc_context* context = thread_context->parent_context;
+  blosc_context* context = thread_context->parent_context;
   int32_t compformat = (*(context->header_flags) & 0xe0) >> 5;
   int dont_split = (*(context->header_flags) & 0x10) >> 4;
   uint8_t blosc_version_format = BLOSC_VERSION_FORMAT;
@@ -606,11 +604,11 @@ static int blosc_c(struct thread_context* thread_context, int32_t blocksize,
   }
 
   /* Shuffle filters */
-  if (*(context->header_flags) & BLOSC_DOSHUFFLE) {
+  if (context->filtercode == BLOSC_SHUFFLE) {
     shuffle(typesize, blocksize, _src, tmp);
     _src = tmp;
   }
-  else if (*(context->header_flags) & BLOSC_DOBITSHUFFLE) {
+  else if (context->filtercode == BLOSC_BITSHUFFLE) {
     bscount = bitshuffle(typesize, blocksize, _src, tmp, tmp2);
     if (bscount < 0)
       return bscount;
@@ -714,9 +712,10 @@ static int blosc_c(struct thread_context* thread_context, int32_t blocksize,
 }
 
 /* Decompress & unshuffle a single block */
-static int blosc_d(struct thread_context* thread_context, int32_t blocksize, int32_t leftoverblock,
-                   const uint8_t* src, uint8_t* dest, int offset, uint8_t* tmp, uint8_t* tmp2) {
-  struct blosc_context* context = thread_context->parent_context;
+static int blosc_d(
+    struct thread_context* thread_context, int32_t blocksize, int32_t leftoverblock,
+    const uint8_t* src, uint8_t* dest, int offset, uint8_t* tmp, uint8_t* tmp2) {
+  blosc_context* context = thread_context->parent_context;
   int32_t compformat = (*(context->header_flags) & 0xe0) >> 5;
   int dont_split = (*(context->header_flags) & 0x10) >> 4;
   uint8_t blosc_version_format = context->src[0];
@@ -730,8 +729,8 @@ static int blosc_d(struct thread_context* thread_context, int32_t blocksize, int
   char* compname;
   int bscount;
 
-  if ((*(context->header_flags) & BLOSC_DOSHUFFLE) || \
-      (*(context->header_flags) & BLOSC_DOBITSHUFFLE)) {
+  if ((context->filtercode == BLOSC_SHUFFLE) || \
+      (context->filtercode == BLOSC_BITSHUFFLE)) {
     _dest = tmp;
   }
 
@@ -803,10 +802,10 @@ static int blosc_d(struct thread_context* thread_context, int32_t blocksize, int
     ntbytes += nbytes;
   } /* Closes j < nsplits */
 
-  if (*(context->header_flags) & BLOSC_DOSHUFFLE) {
+  if (context->filtercode == BLOSC_SHUFFLE) {
     unshuffle(typesize, blocksize, tmp, dest + offset);
   }
-  else if (*(context->header_flags) & BLOSC_DOBITSHUFFLE) {
+  else if (context->filtercode == BLOSC_BITSHUFFLE) {
     bscount = bitunshuffle(typesize, blocksize, tmp, dest + offset, tmp2);
     if (bscount < 0)
       return bscount;
@@ -828,7 +827,7 @@ static int blosc_d(struct thread_context* thread_context, int32_t blocksize, int
 
 /* Serial version for compression/decompression */
 static int serial_blosc(struct thread_context* thread_context) {
-  struct blosc_context* context = thread_context->parent_context;
+  blosc_context* context = thread_context->parent_context;
   int32_t j, bsize, leftoverblock;
   int32_t cbytes;
 
@@ -894,7 +893,7 @@ static int serial_blosc(struct thread_context* thread_context) {
 
 
 /* Threaded version for compression/decompression */
-static int parallel_blosc(struct blosc_context* context) {
+static int parallel_blosc(blosc_context* context) {
   int rc;
 
   /* Set sentinels */
@@ -919,7 +918,7 @@ static int parallel_blosc(struct blosc_context* context) {
 
 
 static struct thread_context* create_thread_context(
-                                  struct blosc_context* context, int32_t tid) {
+                                blosc_context* context, int32_t tid) {
   struct thread_context* thread_context;
   int32_t ebsize;
 
@@ -951,7 +950,7 @@ void free_thread_context(struct thread_context* thread_context) {
 
 /* Do the compression or decompression of the buffer depending on the
    global params. */
-static int do_job(struct blosc_context* context) {
+static int do_job(blosc_context* context) {
   int32_t ntbytes;
 
   /* Run the serial version when nthreads is 1 or when the buffers are
@@ -983,11 +982,9 @@ static int do_job(struct blosc_context* context) {
 		     ((codec) == BLOSC_ZLIB) ||  \
 		     ((codec) == BLOSC_ZSTD) ? 1 : 0 )
 
-static int32_t compute_blocksize(struct blosc_context* context,
-                                 int32_t clevel,
-                                 int32_t typesize,
-                                 int32_t nbytes,
-                                 int32_t forced_blocksize) {
+static int32_t compute_blocksize(
+    blosc_context* context, int32_t clevel, int32_t typesize, int32_t nbytes,
+    int32_t forced_blocksize) {
   int32_t blocksize = nbytes;
 
   /* Protection against very small buffers */
@@ -1055,18 +1052,11 @@ static int32_t compute_blocksize(struct blosc_context* context,
   return blocksize;
 }
 
-static int initialize_context_compression(struct blosc_context* context,
-                                          int clevel,
-                                          int doshuffle,
-                                          size_t typesize,
-                                          size_t sourcesize,
-                                          const void* src,
-                                          void* dest,
-                                          size_t destsize,
-                                          int32_t compressor,
-                                          int32_t blocksize,
-                                          int32_t nthreads,
-                                          schunk_header* schunk) {
+static int initialize_context_compression(
+  blosc_context* context,
+  size_t sourcesize, const void* src, void* dest, size_t destsize, int clevel,
+  int filtercode, size_t typesize, int32_t compressor, int32_t blocksize,
+  int32_t nthreads, schunk_header* schunk) {
 
   /* Set parameters */
   context->compress = 1;
@@ -1076,6 +1066,7 @@ static int initialize_context_compression(struct blosc_context* context,
   context->destsize = (int32_t)destsize;
   context->sourcesize = sourcesize;
   context->typesize = typesize;
+  context->filtercode = filtercode;
   context->compcode = compressor;
   context->nthreads = nthreads;
   context->end_threads = 0;
@@ -1098,8 +1089,10 @@ static int initialize_context_compression(struct blosc_context* context,
   }
 
   /* Shuffle */
-  if (doshuffle != 0 && doshuffle != 1 && doshuffle != 2) {
-    fprintf(stderr, "`shuffle` parameter must be either 0, 1 or 2!\n");
+  if (!((filtercode == BLOSC_NOFILTER) ||
+        (filtercode == BLOSC_SHUFFLE) ||
+        (filtercode == BLOSC_BITSHUFFLE))) {
+    fprintf(stderr, "`filtercode` parameter value `%d` not allowed!\n", filtercode);
     return -10;
   }
 
@@ -1110,22 +1103,32 @@ static int initialize_context_compression(struct blosc_context* context,
   }
 
   /* Get the blocksize */
-  context->blocksize = compute_blocksize(context, clevel, (int32_t)context->typesize, context->sourcesize, blocksize);
+  context->blocksize = compute_blocksize(
+    context, clevel, (int32_t)context->typesize, context->sourcesize, blocksize);
 
   /* Compute number of blocks in buffer */
   context->nblocks = context->sourcesize / context->blocksize;
   context->leftover = context->sourcesize % context->blocksize;
-  context->nblocks = (context->leftover > 0) ? (context->nblocks + 1) : context->nblocks;
+  context->nblocks = (context->leftover > 0) ? \
+   (context->nblocks + 1) : context->nblocks;
 
   return 1;
 }
 
-static int initialize_context_decompression(struct blosc_context* context,
-                                            const void* src,
-                                            void* dest,
-                                            size_t destsize,
-                                            int nthreads,
-                                            schunk_header* schunk) {
+/* Get the filter code from header flags */
+static uint8_t get_filtercode(const uint8_t header_flags) {
+  uint8_t filtercode = BLOSC_NOFILTER;
+
+  if (header_flags & BLOSC_DOSHUFFLE) {
+    filtercode = BLOSC_SHUFFLE;
+  } else if (header_flags & BLOSC_DOBITSHUFFLE) {
+    filtercode = BLOSC_BITSHUFFLE;
+  }
+  return filtercode;
+}
+
+static int initialize_context_decompression(
+    blosc_context* context, const void* src, void* dest, size_t destsize) {
 
   int i;
 
@@ -1134,14 +1137,13 @@ static int initialize_context_decompression(struct blosc_context* context,
   context->dest = (uint8_t*)dest;
   context->destsize = destsize;
   context->num_output_bytes = 0;
-  context->nthreads = nthreads;
   context->end_threads = 0;
-  context->schunk = schunk;
 
   context->header_flags = (uint8_t*)(context->src + 2); /* flags */
   context->typesize = (int32_t)context->src[3];      /* typesize */
   context->sourcesize = sw32_(context->src + 4);     /* buffer size */
   context->blocksize = sw32_(context->src + 8);      /* block size */
+  context->filtercode = get_filtercode(*(context->header_flags));
 
   /* Check that we have enough space to decompress */
   if (context->sourcesize > (int32_t)destsize) {
@@ -1170,7 +1172,7 @@ static int split_block(int compcode, int typesize, int blocksize) {
 	  (blocksize / typesize) >= MIN_BUFFERSIZE);
 }
 
-static int write_compression_header(struct blosc_context* context, int clevel, int doshuffle) {
+static int write_compression_header(blosc_context* context) {
   int32_t compformat;
   int dont_split;
 
@@ -1245,26 +1247,24 @@ static int write_compression_header(struct blosc_context* context, int clevel, i
     *(context->header_flags) |= BLOSC_MEMCPYED;
   }
 
-  if (doshuffle == BLOSC_SHUFFLE) {
+  if (context->filtercode == BLOSC_SHUFFLE) {
     /* Byte-shuffle is active */
     *(context->header_flags) |= BLOSC_DOSHUFFLE;     /* bit 0 set to one in flags */
   }
 
-  if (doshuffle == BLOSC_BITSHUFFLE) {
+  if (context->filtercode == BLOSC_BITSHUFFLE) {
     /* Bit-shuffle is active */
     *(context->header_flags) |= BLOSC_DOBITSHUFFLE;  /* bit 2 set to one in flags */
   }
 
-  dont_split = !split_block(context->compcode, context->typesize,
-			    context->blocksize);
+  dont_split = !split_block(context->compcode, context->typesize, context->blocksize);
   *(context->header_flags) |= dont_split << 4;  /* dont_split is in bit 4 */
-
   *(context->header_flags) |= compformat << 5;  /* compressor format starts at bit 5 */
 
   return 1;
 }
 
-int blosc_compress_context(struct blosc_context* context) {
+int blosc_compress_context(blosc_context* context) {
   int32_t ntbytes = 0;
 
   if (!(*(context->header_flags) & BLOSC_MEMCPYED)) {
@@ -1308,27 +1308,27 @@ int blosc_compress_context(struct blosc_context* context) {
 }
 
 /* The public routine for compression with context. */
-int blosc_compress_ctx(int clevel, int doshuffle, size_t typesize,
-                       size_t nbytes, const void* src, void* dest,
-                       size_t destsize, const char* compressor,
-                       size_t blocksize, int nthreads) {
+int blosc2_compress_ctx(
+    blosc_context* context, size_t nbytes, const void* src, void* dest,
+    size_t destsize) {
   int error, result;
-  struct blosc_context context;
 
-  context.threads_started = 0;
-  error = initialize_context_compression(&context, clevel, doshuffle, typesize, nbytes,
-                                         src, dest, destsize, blosc_compname_to_compcode(compressor),
-                                         blocksize, nthreads, NULL);
-  if (error < 0) { return error; }
-
-  error = write_compression_header(&context, clevel, doshuffle);
-  if (error < 0) { return error; }
-
-  result = blosc_compress_context(&context);
-
-  if (nthreads > 1) {
-    blosc_release_threadpool(&context);
+  if (context->compress != 1) {
+    fprintf(stderr, "Context is not meant for compression.  Giving up.\n");
+    return -10;
   }
+
+  error = initialize_context_compression(
+    context, nbytes, src, dest, destsize,
+    context->clevel, context->filtercode, context->typesize,
+    context->compcode, context->blocksize, context->nthreads,
+    context->schunk);
+  if (error < 0) { return error; }
+
+  error = write_compression_header(context);
+  if (error < 0) { return error; }
+
+  result = blosc_compress_context(context);
 
   return result;
 }
@@ -1342,12 +1342,11 @@ int blosc_compress(int clevel, int doshuffle, size_t typesize, size_t nbytes,
   pthread_mutex_lock(&global_comp_mutex);
 
   error = initialize_context_compression(
-              g_global_context, clevel, doshuffle, typesize, nbytes,
-              src, dest, destsize, g_compressor, g_force_blocksize, g_nthreads,
-              g_schunk);
+    g_global_context, nbytes, src, dest, destsize, clevel, doshuffle, typesize,
+    g_compressor, g_force_blocksize, g_nthreads, g_schunk);
   if (error < 0) { return error; }
 
-  error = write_compression_header(g_global_context, clevel, doshuffle);
+  error = write_compression_header(g_global_context);
   if (error < 0) { return error; }
 
   result = blosc_compress_context(g_global_context);
@@ -1357,20 +1356,16 @@ int blosc_compress(int clevel, int doshuffle, size_t typesize, size_t nbytes,
   return result;
 }
 
-int blosc_run_decompression_with_context(struct blosc_context* context,
-                                         const void* src,
-                                         void* dest,
-                                         size_t destsize,
-                                         int nthreads,
-                                         schunk_header* schunk) {
+int blosc_run_decompression_with_context(
+    blosc_context* context, const void* src, void* dest,
+    size_t destsize) {
   uint8_t version;
   uint8_t versionlz;
   uint32_t ctbytes;
   int32_t ntbytes;
   int error;
 
-  error = initialize_context_decompression(context, src, dest, destsize,
-                                           nthreads, schunk);
+  error = initialize_context_decompression(context, src, dest, destsize);
   if (error < 0) { return error; }
 
   /* Read the header block */
@@ -1411,17 +1406,16 @@ int blosc_run_decompression_with_context(struct blosc_context* context,
 }
 
 /* The public routine for decompression with context. */
-int blosc_decompress_ctx(const void* src, void* dest, size_t destsize,
-                         int nthreads) {
+int blosc2_decompress_ctx(
+    blosc_context* context, const void* src, void* dest, size_t destsize) {
   int result;
-  struct blosc_context context;
 
-  context.threads_started = 0;
-  result = blosc_run_decompression_with_context(&context, src, dest, destsize, nthreads, NULL);
-
-  if (nthreads > 1) {
-    blosc_release_threadpool(&context);
+  if (context->compress != 0) {
+    fprintf(stderr, "Context is not meant for decompression.  Giving up.\n");
+    return -10;
   }
+
+  result = blosc_run_decompression_with_context(context, src, dest, destsize);
 
   return result;
 }
@@ -1429,14 +1423,15 @@ int blosc_decompress_ctx(const void* src, void* dest, size_t destsize,
 
 /* The public routine for decompression.  See blosc.h for docstrings. */
 int blosc_decompress(const void* src, void* dest, size_t destsize) {
-  return blosc_run_decompression_with_context(g_global_context, src, dest, destsize, g_nthreads, g_schunk);
+  /* TODO: Global lock is missing here!! */
+  return blosc_run_decompression_with_context(g_global_context, src, dest, destsize);
 }
-
 
 /* Specific routine optimized for decompression a small number of
    items out of a compressed chunk.  This does not use threads because
    it would affect negatively to performance. */
-int blosc_getitem(const void* src, int start, int nitems, void* dest) {
+int _blosc_getitem(const blosc_context* context, const void* src, int start,
+    int nitems, void* dest) {
   uint8_t* _src = NULL;             /* current pos for source buffer */
   uint8_t version, versionlz;       /* versions for compressed header */
   uint8_t flags;                    /* flags for header */
@@ -1517,38 +1512,77 @@ int blosc_getitem(const void* src, int start, int nitems, void* dest) {
       cbytes = bsize2;
     }
     else {
-      struct blosc_context context;
-      uint8_t* tmp;
-      uint8_t* tmp2;
-      uint8_t* tmp3;
-      /* blosc_d only uses typesize, blocksize, flags and schunk info */
-      context.typesize = typesize;
-      context.blocksize = blocksize;
-      context.header_flags = &flags;
-      context.schunk = g_schunk;
-      context.serial_context = create_thread_context(&context, 0);
-      tmp = context.serial_context->tmp;
-      tmp2 = context.serial_context->tmp2;
-      tmp3 = context.serial_context->tmp3;
+      struct thread_context* scontext = context->serial_context;
+
+      /* Resize the temporaries in serial context if needed */
+      if (blocksize != scontext->tmpblocksize) {
+        my_free(scontext->tmp);
+        scontext->tmp = my_malloc(blocksize + ebsize + blocksize);
+        scontext->tmp2 = scontext->tmp + blocksize;
+        scontext->tmp3 = scontext->tmp + blocksize + ebsize;
+        scontext->tmpblocksize = blocksize;
+      }
 
       /* Regular decompression.  Put results in tmp2. */
-      cbytes = blosc_d(context.serial_context, bsize, leftoverblock,
+      cbytes = blosc_d(context->serial_context, bsize, leftoverblock,
                        (uint8_t*)src + sw32_(bstarts + j * 4),
-                       tmp2, 0, tmp, tmp3);
+                       scontext->tmp2, 0, scontext->tmp, scontext->tmp3);
       if (cbytes < 0) {
         ntbytes = cbytes;
         break;
       }
       /* Copy to destination */
-      memcpy((uint8_t*)dest + ntbytes, tmp2 + startb, bsize2);
+      memcpy((uint8_t*)dest + ntbytes, scontext->tmp2 + startb, bsize2);
       cbytes = bsize2;
-
-      free_thread_context(context.serial_context);
     }
     ntbytes += cbytes;
   }
 
   return ntbytes;
+}
+
+
+/* Specific routine optimized for decompression a small number of
+   items out of a compressed chunk.  Public non-contextual API. */
+int blosc_getitem(const void* src, int start, int nitems, void* dest) {
+  uint8_t* _src = (uint8_t*)(src);
+  blosc_context context;
+  int result;
+
+  /* Minimally populate the context */
+  context.typesize = (int32_t)_src[3];
+  context.blocksize = sw32_(_src + 8);
+  context.header_flags = _src + 2;
+  context.filtercode = get_filtercode(*(_src + 2));
+  context.schunk = g_schunk;
+  context.serial_context = create_thread_context(&context, 0);
+
+  /* Call the actual getitem function */
+  result = _blosc_getitem(&context, src, start, nitems, dest);
+
+  /* Release resources */
+  free_thread_context(context.serial_context);
+  return result;
+}
+
+int blosc2_getitem_ctx(blosc_context* context, const void* src, int start,
+    int nitems, void* dest) {
+  uint8_t* _src = (uint8_t*)(src);
+  int result;
+
+  /* Minimally populate the context */
+  context->typesize = (int32_t)_src[3];
+  context->blocksize = sw32_(_src + 8);
+  context->header_flags = _src + 2;
+  context->filtercode = get_filtercode(*(_src + 2));
+  if (context->serial_context == NULL) {
+    context->serial_context = create_thread_context(context, 0);
+  }
+
+  /* Call the actual getitem function */
+  result = _blosc_getitem(context, src, start, nitems, dest);
+
+  return result;
 }
 
 
@@ -1732,7 +1766,7 @@ static void* t_blosc(void* ctxt) {
   return (NULL);
 }
 
-static int init_threads(struct blosc_context* context) {
+static int init_threads(blosc_context* context) {
   int32_t tid;
   int rc2;
   struct thread_context* thread_context;
@@ -1790,12 +1824,13 @@ int blosc_set_nthreads(int nthreads_new) {
     blosc_destroy();
     blosc_init();
     g_nthreads = nthreads_new;
+    g_global_context->nthreads = nthreads_new;
   }
 
   return ret;
 }
 
-int blosc_set_nthreads_(struct blosc_context* context) {
+int blosc_set_nthreads_(blosc_context* context) {
   if (context->nthreads <= 0) {
     fprintf(stderr, "Error.  nthreads must be a positive integer");
     return -1;
@@ -1980,10 +2015,11 @@ void blosc_set_blocksize(size_t size) {
    reachable (the default). */
 void blosc_set_schunk(schunk_header* schunk) {
   g_schunk = schunk;
+  g_global_context->schunk = schunk;
 }
 
-struct blosc_context* create_context(int nthreads) {
-  struct blosc_context* context = (struct blosc_context*)my_malloc(sizeof(struct blosc_context));
+blosc_context* create_context(int nthreads) {
+  blosc_context* context = (blosc_context*)my_malloc(sizeof(blosc_context));
 
   /* Initialize some struct components */
   context->serial_context = NULL;
@@ -2003,13 +2039,13 @@ void blosc_destroy(void) {
   g_initlib = 0;
   blosc_release_threadpool(g_global_context);
   if (g_global_context->serial_context != NULL) {
-    my_free(g_global_context->serial_context);
+    free_thread_context(g_global_context->serial_context);
   }
   my_free(g_global_context);
   pthread_mutex_destroy(&global_comp_mutex);
 }
 
-int blosc_release_threadpool(struct blosc_context* context) {
+int blosc_release_threadpool(blosc_context* context) {
   int32_t t;
   void* status;
   int rc;
@@ -2059,4 +2095,50 @@ int blosc_release_threadpool(struct blosc_context* context) {
 
 int blosc_free_resources(void) {
   return blosc_release_threadpool(g_global_context);
+}
+
+
+/* Contexts */
+
+/* Create a context for compression */
+blosc_context* blosc2_create_cctx(blosc2_context_cparams* cparams) {
+  int error;
+  blosc_context* context = (blosc_context*)my_malloc(sizeof(blosc_context));
+  memset(context, 0, sizeof(blosc_context));
+
+  context->compress = 1;   /* meant for compression */
+  /* Populate the context, using default values for zeroed values */
+  context->typesize = cparams->typesize ? cparams->typesize : 8;
+  context->compcode = cparams->compcode ? cparams->compcode : BLOSC_BLOSCLZ;
+  context->clevel = cparams->clevel ? cparams->clevel : 5;
+  context->filtercode = cparams->filtercode ? cparams->filtercode : BLOSC_SHUFFLE;
+  context->blocksize = cparams->blocksize ? cparams->blocksize : 0;
+  context->nthreads = cparams->nthreads ? cparams->nthreads : 1;
+  context->schunk = cparams->schunk ? cparams->schunk : NULL;
+
+  return context;
+}
+
+
+
+/* Create a context for decompression */
+blosc_context* blosc2_create_dctx(blosc2_context_dparams* dparams) {
+  int error;
+  blosc_context* context = (blosc_context*)my_malloc(sizeof(blosc_context));
+  memset(context, 0, sizeof(blosc_context));
+
+  context->compress = 0;   /* Meant for decompression */
+  /* Populate the context, using default values for zeroed values */
+  context->nthreads = dparams->nthreads ? dparams->nthreads : 1;
+  context->schunk = dparams->schunk ? dparams->schunk : NULL;
+
+  return context;
+}
+
+void blosc2_free_ctx(blosc_context* context) {
+  blosc_release_threadpool(context);
+  if (context->serial_context != NULL) {
+    free_thread_context(context->serial_context);
+  }
+  my_free(context);
 }