+++ examples/contexts.c
@@ -7,13 +7,13 @@
 
     To compile this program:
 
-    $ gcc -O simple.c -o simple -lblosc
+    $ gcc -O contexts.c -o contexts -lblosc
 
     To run:
 
-    $ ./simple
+    $ ./contexts
     Blosc version info: 2.0.0a2 ($Date:: 2016-01-08 #$)
-    Compression: 40000000 -> 832774 (48.0x)
+    Compression: 40000000 -> 999393 (40.0x)
     Correctly extracted 5 elements from compressed chunk!
     Decompression succesful!
     Succesful roundtrip!
@@ -36,7 +36,11 @@ int main() {
   int isize = SIZE * sizeof(float), osize = SIZE * sizeof(float);
   int dsize = SIZE * sizeof(float), csize;
   int i, ret;
+  blosc2_context_cparams cparams = BLOSC_CPARAMS_DEFAULTS;
+  blosc2_context_dparams dparams = BLOSC_DPARAMS_DEFAULTS;
+  blosc_context *cctx, *dctx;
 
+  /* Initialize dataset */
   for (i = 0; i < SIZE; i++) {
     data[i] = i;
   }
@@ -44,13 +48,16 @@ int main() {
   printf("Blosc version info: %s (%s)\n",
          BLOSC_VERSION_STRING, BLOSC_VERSION_DATE);
 
-  /* Initialize the Blosc compressor */
-  blosc_init();
-  blosc_set_nthreads(NTHREADS);
+  /* Create a context for compression */
+  cparams.typesize = sizeof(float);
+  cparams.compcode = BLOSC_BLOSCLZ;
+  cparams.filtercode = BLOSC_SHUFFLE;
+  cparams.clevel = 5;
+  cparams.nthreads = NTHREADS;
+  cctx = blosc2_create_cctx(&cparams);
 
   /* Compress with clevel=5 and shuffle active  */
-  csize = blosc_compress(5, BLOSC_BITSHUFFLE, sizeof(float), isize, data,
-                         data_out, osize);
+  csize = blosc2_compress_ctx(cctx, isize, data, data_out, osize);
   if (csize == 0) {
     printf("Buffer is uncompressible.  Giving up.\n");
     return 1;
@@ -62,22 +69,26 @@ int main() {
 
   printf("Compression: %d -> %d (%.1fx)\n", isize, csize, (1. * isize) / csize);
 
-  ret = blosc_getitem(data_out, 5, 5, data_subset);
+  /* Create a context for decompression */
+  dparams.nthreads = NTHREADS;
+  dctx = blosc2_create_dctx(&dparams);
+
+  ret = blosc2_getitem_ctx(dctx, data_out, 5, 5, data_subset);
   if (ret < 0) {
-    printf("Error in blosc_getitem().  Giving up.\n");
+    printf("Error in blosc2_getitem_ctx().  Giving up.\n");
     return 1;
   }
 
   for (i = 0; i < 5; i++) {
     if (data_subset[i] != data_subset_ref[i]) {
-      printf("blosc_getitem() fetched data differs from original!\n");
+      printf("blosc2_getitem_ctx() fetched data differs from original!\n");
       return -1;
     }
   }
   printf("Correctly extracted 5 elements from compressed chunk!\n");
 
   /* Decompress  */
-  dsize = blosc_decompress(data_out, data_dest, dsize);
+  dsize = blosc2_decompress_ctx(dctx, data_out, data_dest, dsize);
   if (dsize < 0) {
     printf("Decompression error.  Error code: %d\n", dsize);
     return dsize;
@@ -85,16 +96,17 @@ int main() {
 
   printf("Decompression succesful!\n");
 
-  /* After using it, destroy the Blosc environment */
-  blosc_destroy();
-
   for (i = 0; i < SIZE; i++) {
     if (data[i] != data_dest[i]) {
       printf("Decompressed data differs from original!\n");
       return -1;
     }
   }
-
   printf("Succesful roundtrip!\n");
+
+  /* Release resources */
+  blosc2_free_ctx(cctx);
+  blosc2_free_ctx(dctx);
+
   return 0;
 }